+++ components/menu/Menu.vue
@@ -38,8 +38,7 @@ const props = withDefaults(defineProps<Props>(), {
 const emits = defineEmits(['update:openKeys', 'update:selectedKeys', 'click', 'select', 'openChange'])
 </script>
 <template>
-  <div class="m-menu">
-  </div>
+  <div class="m-menu"> </div>
 </template>
 <style lang="less" scoped>
 .m-menu {