+++ src/views/CascaderView.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import { ref, watchEffect } from 'vue'
-import { generate } from '@ant-design/colors'
 const options = ref([
   {
     value: '1',
@@ -186,21 +185,9 @@ const sizeOptions = [
 ]
 const size = ref('large')
 const selectedValue = ref(['2', '21', '212'])
-const primaryColor = ref('#ff6900')
-const primaryShadowColor = ref('rgba(255, 116, 32, 0.1)')
 watchEffect(() => {
   console.log('selectedValue', selectedValue.value)
 })
-function getThemeStyle(color: string) {
-  const colorPalettes = generate(color)
-  const style = {
-    '--select-primary-color-hover': colorPalettes[4],
-    '--select-primary-color-focus': colorPalettes[4],
-    '--select-primary-shadow-color': primaryShadowColor.value,
-    '--select-item-bg-color-active': colorPalettes[0]
-  }
-  return style
-}
 function onChange(values: (number | string)[], labels: string[]) {
   console.log('values', values)
   console.log('labels', labels)
@@ -249,20 +236,5 @@ function filter(inputValue: string, option: any) {
     <Cascader :options="options" v-model="selectedValue" search @change="onChange" />
     <h2 class="mt30 mb10">自定义搜索过滤函数</h2>
     <Cascader :options="options" v-model="selectedValue" allow-clear search :filter="filter" @change="onChange" />
-    <h2 class="mt30 mb10">自定义主题色</h2>
-    <Space vertical>
-      <Space align="center"> primaryColor:<ColorPicker style="width: 200px" v-model:value="primaryColor" /> </Space>
-      <Space align="center">
-        primaryShadowColor:<ColorPicker style="width: 200px" v-model:value="primaryShadowColor" />
-      </Space>
-      <Cascader
-        :style="getThemeStyle(primaryColor)"
-        :options="options"
-        v-model="selectedValue"
-        allow-clear
-        search
-        @change="onChange"
-      />
-    </Space>
   </div>
 </template>